Create StringRemoveWhiteSpace.java
String API in Java provides several utility methods to remove white space from String in Java, from the beginning, end, and between words. White space means blank space including tab characters. One of the classic methods to remove white space from the beginning and end of a string in Java is the trim() method, which removes white space from beginning and end. While using the trim() method, the key thing to remember is that trim() returns a different String object than the older one because String is immutable in Java.